修复拖拽时卡顿的问题
