--- HuiDesktop/BasicWindow.xaml
@@ -1,10 +1,0 @@
-﻿<Window x:Class="HuiDesktop.BasicWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:HuiDesktop"
-        mc:Ignorable="d"
-        Title="HuiDesktop" WindowStyle="None" AllowsTransparency="True" Background="Transparent" SizeChanged="Window_SizeChanged" Icon="Resources/GlobalIcon.ico" Closing="Window_Closing">
-</Window>
-
